api/requirements.txt
Django==1.10.6
djangorestframework==3.6.2
django-rest-swagger==2.1.2
django-extensions==1.7.7
requests==2.13.0
urllib3==1.20
django-environ==0.4.1
jsonschema==3.0.2
drf-nested-routers==0.90.0
